requirements.txt
fastapi
uvicorn
python-multipart
filelock
schedule
psutil
pydantic-settings
